backend/requirements.txt
anthropic==0.40.0
fastapi==0.115.0
uvicorn[standard]==0.32.0
supabase==2.9.0
python-dotenv==1.0.1
httpx==0.28.0
pydantic==2.10.0
python-multipart==0.0.19
